requirements.txt
fastapi
uvicorn
websockets
sqlmodel
aiosqlite
strands-agents
strands-agents-tools
python-dotenv
jinja2